chrono: convert to double precision for overflow purposes
# Conflicts:
#	ajek-framework/h/x-chrono.h

+++ ajek-framework/src/x-stdlib.cpp
@@ -407,7 +407,7 @@ __exi xString xFixFilenameForPlatform(const xString& src)
 // x-chrono.h things  (not in a mood to put these in their own module, yet...)
 
 host_tick_t		HostClockTick::s_ticks_per_second;
-float			HostClockTick::s_ticks_per_second_f;
+double			HostClockTick::s_ticks_per_second_f;
 
 void HostClockTick::Init()
 {
